# BilfenAppPhp.RestAPI/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bilfen API",
        "version": "v1"
    },
    "components": {
        "schemas": {
...
            "Country": {
                "type": "object",
                "required": ["Id", "Name", "createdAt", "updatedAt"],
                "properties": {
                    "Id": { "type": "string", "description": "The id of the country" },
                    "Name": { "type": "string", "description": "The country's name" },
                    "createdAt": { "type": "string", "description": "The country's created date" },
                    "updatedAt": { "type": "string", "description": "The country's updated date" }
                },
                "example": {
                    "Id": "string",
                    "Name": "string",
                    "createdAt": "string",
                    "updatedAt": "string"
                }
            },
...
            "CreateUpdateCountry": {
                "type": "object",
                "required": ["Id", "Name"],
                "properties": {
                    "Id": { "type": "string", "description": "The id of the country" },
                    "Name": { "type": "string", "description": "The country's name" }
                },
                "example": {
                    "Id": "string",
                    "Name": "string"
                }
            },
...
        }
    },
    "paths": {
...
        "/api/countries/GetCountry/{id}": {
            "get": {
                "tags": [
                    "Countries"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "id",
                    "description": "Country Id",
                    "schema": {
                        "type": "string"
                    },
                    "required": true
                }],
                "operationId": "getCountry",
                "responses": {
                    "200": {
                        "description": "Retrives a country by id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Country"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/countries/CreateCountry": {
            "post": {
                "tags": ["Countries"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateCountry"
                            }
                        }
                    }
                },
                "operationId": "createCountry",
                "responses": {
                    "200": {
                        "description": "Create a new country data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Country"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/countries/EditCountry": {
            "put": {
                "tags": ["Countries"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateCountry"
                            }
                        }
                    }
                },
                "operationId": "editCountry",
                "responses": {
                    "200": {
                        "description": "Edit an existing country data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Country"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/countries/DeleteCountry/{id}": {
            "delete": {
                "tags": ["Countries"],
                "parameters": [{
                    "in": "path",
                    "name": "id",
                    "description": "Country Id",
                    "schema": {
                        "type": "string"
                    },
                    "required": true
                }],
                "responses": {
                    "204": {
                        "description": "Delete an existing country data"
                    }
                }
            }
        },
...
    },
...
}